Release 0.1.8

+++ common/stack.go
@@ -2,6 +2,7 @@ package common
 
 import (
 	"bytes"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"github.com/aws/aws-sdk-go/aws"
@@ -12,13 +13,31 @@ import (
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/aws/aws-sdk-go/service/ec2/ec2iface"
 	"io"
+	"io/ioutil"
+	"os"
 	"strings"
 	"time"
+	"github.com/briandowns/spinner"
 )
 
 // CreateStackName will create a name for a stack
 func CreateStackName(stackType StackType, names ...string) string {
 	return fmt.Sprintf("mu-%s-%s", stackType, strings.Join(names, "-"))
+}
+
+// GetStackOverrides will get the overrides from the config
+func GetStackOverrides(stackName string) interface{} {
+	if stackOverrides == nil {
+		return nil
+	}
+
+	return stackOverrides[stackName]
+}
+
+var stackOverrides map[string]interface{}
+
+func registerStackOverrides(overrides map[string]interface{}) {
+	stackOverrides = overrides
 }
 
 // StackWaiter for waiting on stack status to be final
@@ -62,12 +81,13 @@ type StackManager interface {
 }
 
 type cloudformationStackManager struct {
+	dryrun bool
 	cfnAPI cloudformationiface.CloudFormationAPI
 	ec2API ec2iface.EC2API
 }
 
 // NewStackManager creates a new StackManager backed by cloudformation
-func newStackManager(sess *session.Session) (StackManager, error) {
+func newStackManager(sess *session.Session, dryrun bool) (StackManager, error) {
 	log.Debug("Connecting to CloudFormation service")
 	cfnAPI := cloudformation.New(sess)
 
@@ -75,6 +95,7 @@ func newStackManager(sess *session.Session) (StackManager, error) {
 	ec2API := ec2.New(sess)
 
 	return &cloudformationStackManager{
+		dryrun: dryrun,
 		cfnAPI: cfnAPI,
 		ec2API: ec2API,
 	}, nil
@@ -117,6 +138,16 @@ func buildStackTags(tags map[string]string) []*cloudformation.Tag {
 func (cfnMgr *cloudformationStackManager) UpsertStack(stackName string, templateBodyReader io.Reader, parameters map[string]string, tags map[string]string) error {
 	stack := cfnMgr.AwaitFinalStatus(stackName)
 
+	// delete stack if in rollback status
+	if stack != nil && stack.Status == cloudformation.StackStatusRollbackComplete {
+		log.Warningf("  Stack '%s' was in '%s' status, deleting...", stackName, stack.Status)
+		err := cfnMgr.DeleteStack(stackName)
+		if err != nil {
+			return err
+		}
+		stack = cfnMgr.AwaitFinalStatus(stackName)
+	}
+
 	// load the template
 	templateBodyBytes := new(bytes.Buffer)
 	templateBodyBytes.ReadFrom(templateBodyReader)
@@ -127,6 +158,9 @@ func (cfnMgr *cloudformationStackManager) UpsertStack(stackName string, template
 
 	// stack tags
 	stackTags := buildStackTags(tags)
+
+	// directory to write cfn to
+	cfnDirectory := fmt.Sprintf("%s/mu-cloudformation", os.TempDir())
 
 	cfnAPI := cfnMgr.cfnAPI
 	if stack == nil || stack.Status == "" {
@@ -143,6 +177,13 @@ func (cfnMgr *cloudformationStackManager) UpsertStack(stackName string, template
 			TemplateBody: templateBody,
 			Tags:         stackTags,
 		}
+
+		if cfnMgr.dryrun {
+			writeTemplateAndConfig(cfnDirectory, stackName, templateBodyBytes, parameters)
+			log.Infof("  DRYRUN: Skipping create of stack named '%s'.  Template and parameters written to '%s'", stackName, cfnDirectory)
+			return nil
+		}
+
 		_, err := cfnAPI.CreateStack(params)
 		log.Debug("  Create stack complete err=%s", err)
 		if err != nil {
@@ -172,6 +213,12 @@ func (cfnMgr *cloudformationStackManager) UpsertStack(stackName string, template
 			Tags: stackTags,
 		}
 
+		if cfnMgr.dryrun {
+			writeTemplateAndConfig(cfnDirectory, stackName, templateBodyBytes, parameters)
+			log.Infof("  DRYRUN: Skipping update of stack named '%s'.  Template and parameters written to '%s'", stackName, cfnDirectory)
+			return nil
+		}
+
 		_, err := cfnAPI.UpdateStack(params)
 		log.Debug("  Update stack complete err=%s", err)
 		if err != nil {
@@ -195,6 +242,12 @@ func (cfnMgr *cloudformationStackManager) AwaitFinalStatus(stackName string) *St
 	params := &cloudformation.DescribeStacksInput{
 		StackName: aws.String(stackName),
 	}
+
+	// initialize Spinner
+	spinner := spinner.New(spinner.CharSets[9], 100*time.Millisecond)
+	spinner.Start()
+	defer spinner.Stop()
+
 	resp, err := cfnAPI.DescribeStacks(params)
 
 	if err == nil && resp != nil && len(resp.Stacks) == 1 {
@@ -370,8 +423,32 @@ func (cfnMgr *cloudformationStackManager) DeleteStack(stackName string) error {
 
 	params := &cloudformation.DeleteStackInput{StackName: aws.String(stackName)}
 
+	if cfnMgr.dryrun {
+		log.Infof("  DRYRUN: Skipping delete of stack named '%s'", stackName)
+		return nil
+	}
+
 	log.Debugf("Deleting stack named '%s'", stackName)
 
 	_, err := cfnAPI.DeleteStack(params)
 	return err
+}
+
+func writeTemplateAndConfig(cfnDirectory string, stackName string, templateBodyBytes *bytes.Buffer, parameters map[string]string) error {
+	os.MkdirAll(cfnDirectory, 0700)
+	templateFile := fmt.Sprintf("%s/template-%s.yml", cfnDirectory, stackName)
+	err := ioutil.WriteFile(templateFile, templateBodyBytes.Bytes(), 0600)
+	if err != nil {
+		return err
+	}
+
+	configMap := make(map[string]map[string]string)
+	configMap["Parameters"] = parameters
+	configBody, err := json.MarshalIndent(configMap, "", "  ")
+	if err != nil {
+		return err
+	}
+	configFile := fmt.Sprintf("%s/config-%s.json", cfnDirectory, stackName)
+	err = ioutil.WriteFile(configFile, configBody, 0600)
+	return nil
 }
